flaskapp/requirements.txt
Flask
PyMySQL
python-dateutil